% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readVcf}
\alias{readVcf}
\title{Read VCF compressed with gzip/bgzip}
\usage{
readVcf(filename)
}
\arguments{
\item{filename}{A file name of a compressed VCF file ending with .gz or .bgz}
}
\value{
A integer genotype matrix of a size [# variants] x [# samples]. 
        Only GT field will be recognized. The row names will be variant IDs, 
        in the format of [CHROM]:[POS]:[REF]:[ALT], and the column names 
        will be the sample IDs
}
\description{
This function reads a compressed gzip/bgzip files and returns 
a genotype matrix of size [# variants] x [# samples].
}
\examples{
fpath = system.file("extdata",
  "1000G_exome_chr20_example_softFiltered.calls.vcf.gz",
  package="tinyVcfReader")
mat = readVcf(fpath)
dim(mat)
}
